rename storage with consent and move to lib
